# Fix journal and tag table indexing