## Changes

* Change table creation to match old Sql persistence tables
  * Change journal table primary key from composite (persistence_id, sequence_nr) to ordering
  * Add custom unique (persistence_id, sequence_nr) constraint on journal table
  * Add custom persistence_id index to journal table
  * Add custom sequence_nr index to journal table

* Change tag table primary key to composite (ordering, tag)
* Add custom (persistence_id, sequence_nr) index to tag table (used for deletion)
* Add custom tag index on tag table (used for tag query lookup)

### Latest `dev` Benchmarks 

Tested on 3 million events database with 10 tagged events at the end of the table.

failed

### This PR's Benchmarks

Tested on 3 million events database with 10 tagged events at the end of the table.

|     Method |  TagMode |         Mean |      Error |     StdDev |    Gen0 | Allocated |
|----------- |--------- |-------------:|-----------:|-----------:|--------:|----------:|
| QueryByTag |      Csv | 1,745.617 ms | 27.1483 ms | 24.0662 ms |       - | 398.95 KB |
| QueryByTag | TagTable |     3.000 ms |  0.0589 ms |  0.0917 ms | 35.1563 | 301.36 KB |

Note that there are sone Gen0 allocation with tag table because we need to perform the query in 2 steps, the join and then retrieving the tags.

| Test            | SqlServer | SqlServer</br>Batching | Old</br>Indexing | vs Normal | vs Batching | New</br>Indexing |  vs Old |
|:----------------|----------:|-------------------:|-------------:|----------:|------------:|-------------:|--------:|
| Persist         |       304 |                299 |          496 |   163.16% |     165.89% |          477 |  96.17% |
| PersistAll      |      1139 |               1275 |         7893 |   692.98% |     619.06% |         7196 |  91.17% |
| PersistAsync    |      1021 |               1371 |        31813 |  3115.87% |    2320.42% |        26956 |  84.73% |
| PersistAllAsync |      2828 |               1395 |        29634 |  1047.88% |    2124.30% |        27819 |  93.88% |
| PersistGroup10  |       986 |               1034 |         1675 |   169.88% |     161.99% |         1779 | 106.21% |
| PersistGroup25  |      1034 |               1010 |         3054 |   295.36% |     302.38% |         3201 | 104.81% |
| PersistGroup50  |       971 |                980 |         4932 |   507.93% |     503.27% |         5021 | 101.80% |
| PersistGroup100 |      1054 |               1304 |         6249 |   592.88% |     479.22% |         6965 | 111.46% |
| PersistGroup200 |       990 |               1662 |         8086 |   816.77% |     486.52% |         9100 | 112.54% |
| PersistGroup400 |      1049 |               2113 |         7237 |   689.59% |     342.50% |         9224 | 127.46% |
| Recovering      |     60516 |              77688 |        64457 |   106.51% |      82.96% |        62538 |  97.02% |
| RecoveringTwo   |     60730 |              53062 |        43325 |    71.34% |      81.65% |        68427 | 157.94% |
| RecoveringFour  |     86107 |              73218 |        66512 |    77.24% |      90.84% |       108357 | 162.91% |
| Recovering8     |    116401 |             101549 |       103463 |    88.89% |     101.88% |       130808 | 126.43% |

Note that, while there's a write penalty due to extra indices, write and read performance actually improves with parallelization